% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDeltanetsPlusObj.R
\name{createDeltanetsPlusObj}
\alias{createDeltanetsPlusObj}
\title{creating a deltanetsPlus object.}
\usage{
createDeltanetsPlusObj(
  lfc = lfc,
  pval = NULL,
  tp = NULL,
  experiment = NULL,
  p.thres = 0.05
)
}
\arguments{
\item{lfc}{The numeric matrix or data.frame of log2FC data. Each row represents a gene and each column represents a sample.}

\item{pval}{p-values for log2FC values. If pval==NULL, no filtering of log2FC data.}

\item{tp}{A vector of time points. The length of the vector should be the same as the number of samples. If steady-state samples, set the corresponding elements of tp to zero.}

\item{experiment}{A numeric vector for indicating the sampe experiment/perturbation. The (time-series) samples from the same experiment/perturbation should have the same unique index. The length of the vector should be the same as the number of samples.}

\item{threshold}{A threshold value for \code{pval}.}
}
\value{
a deltanetsPlus object
}
\description{
This function creastes a deltanetsPlus object, which is an input format of \code{deltanetsPlus} function.
}
