% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankp.R
\name{rankp}
\alias{rankp}
\title{rank genes from P matrix}
\usage{
rankp(scoremat, glist = NULL)
}
\arguments{
\item{scoremat}{a gene score matrix returned by \code{deltanetsPlus} function.}

\item{glist}{The list of genes corresponding to the rows of P matrix. If glist is not given, gene symbols are just numbers (e.g. G1, G2, G3, ...)}
}
\value{
a list of rankOfGenes and rankedGenes
\item{rankOfGenes}{The matrix of gene ranks. Rows and columns correspond to genes and samples in the same order as the one in P matrix, respectively.A maximum rank is assigned for the genes with zero values in the P matrix}
\item{rankedGenes}{A list of ranked gene lists. Each element in the list includes a ranked gene list for the corresponding sample. The genes closer to the front are the gene targets with the higher confidence. }
}
\description{
The function for ranking genes per sample, based on absolute magnitudes in estimates of perturbation impacts
}
