% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsig2interp.R
\name{unsig2interp}
\alias{unsig2interp}
\title{interpolating log2FC values with p-value > threshold}
\usage{
unsig2interp(lfc, lfc.pval, group, tp, threshold)
}
\arguments{
\item{lfc}{log2FC data}

\item{lfc.pval}{p-values for log2FC values, e.g. which can be obtained when averaging replicates using limma.}

\item{group}{A vector of indices indicating the set of samples from a particular drug/compound treatment. The (time-series) samples from the same drug treatment experiment should have the same unique index. The length of the vector should be the same as the number of samples.}

\item{tp}{A vector of time points of the samples in the matrix lfc. The length of the vector should be the same as the number of samples (i.e. the number of columns in the matrix \code{lfc}).}

\item{threshold}{A threshold value for \code{lfc.pval}.}
}
\value{
\item{lfc.new}
}
\description{
The function for choosing log2FC values with p-value < threshold and interpolating the values for the time points between them.
}
